♻️ refactor: 重构 openai 接口调用逻辑，将插件 schema 开启关闭逻辑与接口解耦
