+++ src/pages/api/openai.ts
@@ -1,10 +1,8 @@
 import { OpenAIStream, OpenAIStreamCallbacks, StreamingTextResponse } from 'ai';
 import { Configuration, OpenAIApi } from 'openai-edge';
-import { ChatCompletionFunctions } from 'openai-edge/types/api';
 
 import { getServerConfig } from '@/config/server';
 import { createErrorResponse } from '@/pages/api/error';
-import { PluginsMap } from '@/plugins';
 import { ErrorType } from '@/types/fetch';
 import { OpenAIStreamPayload } from '@/types/openai';
 
@@ -38,32 +36,18 @@ export const createChatCompletion = async ({
 
   const openai = createOpenAI(OPENAI_API_KEY, endpoint);
 
-  const { messages, plugins: enabledPlugins, ...params } = payload;
+  // ============  1. 前置处理 messages   ============ //
+  const { messages, ...params } = payload;
 
-  // ============  1. 前置处理 functions   ============ //
-
-  const filterFunctions: ChatCompletionFunctions[] = Object.values(PluginsMap)
-    .filter((p) => {
-      // 如果不存在 enabledPlugins，那么全部不启用
-      if (!enabledPlugins) return false;
-
-      // 如果存在 enabledPlugins，那么只启用 enabledPlugins 中的插件
-      return enabledPlugins.includes(p.name);
-    })
-    .map((f) => f.schema);
-
-  const functions = filterFunctions.length === 0 ? undefined : filterFunctions;
-
-  // ============  2. 前置处理 messages   ============ //
   const formatMessages = messages.map((m) => ({
     content: m.content,
     name: m.name,
     role: m.role,
   }));
 
-  // ============  3. 发送请求   ============ //
+  // ============  2. 发送请求   ============ //
 
-  const requestParams = { functions, messages: formatMessages, stream: true, ...params };
+  const requestParams = { messages: formatMessages, stream: true, ...params };
 
   let response: Response;
